[refactor] #230 - PR 전 코드 수정
